requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
audioread==3.0.0
auditok==0.1.5
Automat==0.8.0
autopep8==2.0.1
# av==10.0.0
# ffmpeg-dev
blinker==1.4
boxsdk==3.6.1
cachetools==5.2.1
cchardet==2.1.7
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
# cloud-init==22.4.2
colorama==0.4.3
# command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
# distro-info===0.23ubuntu1
docker==6.0.1
entrypoints==0.3
fastapi==0.89.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
ffsubsync==0.4.23
filelock==3.9.0
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
gast==0.4.0
Glances==3.3.1
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
gradio==3.16.2
grpcio==1.51.1
h11==0.12.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==0.13.7
httplib2==0.21.0
httpx==0.19.0
huggingface-hub==0.11.1
hyperframe==6.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.24.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
incremental==16.10.1
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.4
# language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==15.0.6.1
librosa==0.8.1
linkify-it-py==1.0.3
# llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==2.1.0
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==4.2.0
moviepy==1.0.3
multidict==6.0.4
netifaces==0.10.4
nltk==3.8.1
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
opt-einsum==3.3.0
orjson==3.8.5
packaging==23.0
pandas==1.5.2
pexpect==4.6.0
pilgram==1.2.1
Pillow==9.4.0
pipenv==11.9.0
pooch==1.6.0
proglog==0.1.10
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
PyDrive==1.3.1
pydub==0.25.1
Pygments==2.14.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
# pyparted==3.11.2
pyrsistent==0.15.5
pyserial==3.4
pysrt==1.1.2
pysubs2==1.6.0
# python-apt
python-dateutil==2.8.2
# python-debian===0.1.36ubuntu1
python-multipart==0.0.5
pytube==12.1.2
pytz==2022.7.1
pyxattr==0.6.1
PyYAML==5.3.1
regex==2022.10.31
replicate==0.4.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
resampy==0.4.2
rfc3986==1.5.0
rich==13.1.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
# sos==4.4
soundfile==0.11.0
spleeter==2.3.2
srt==3.5.2
ssh-import-id==5.10
stable-ts==1.0.2
starlette==0.22.0
# systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
transformers==4.25.1
Twisted==18.9.0
typer==0.3.2
typing-extensions==4.4.0
# ubuntu-advantage-tools==27.12
uc-micro-py==1.0.1
# ufw==0.36
ujson==5.7.0
# unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
# WALinuxAgent==2.2.46
webrtcvad==2.0.10
websocket-client==1.4.2
websockets==10.4
whisper==1.0
wrapt==1.14.1
yarl==1.8.2
youtube-dl
zipp==1.0.0
# zope.interface==4.7.1
